# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  bigbucks_solution_auth_rest-api_controllers_types.Role:
    properties:
      description:
        type: string
      extraAttrs:
        additionalProperties: true
        type: object
      name:
        type: string
    required:
    - name
    type: object
...
  types.AuthorizeResponse:
    properties:
      status:
        type: boolean
    type: object
  types.CheckPermissionBody:
    properties:
      action:
        type: string
      orgID:
        type: integer
      resource:
        type: string
      scope:
        type: string
    type: object
  types.CreatePermissionBody:
    properties:
      action:
        minLength: 3
        type: string
      resource:
        minLength: 3
        type: string
      scope:
        minLength: 3
        type: string
    required:
    - action
    - resource
    - scope
    type: object
  types.ListRolePermission:
    properties:
      action:
        type: string
      isHidden:
        type: boolean
      isLocked:
        type: boolean
      resource:
        type: string
      scope:
        type: string
    type: object
  types.ListRoleResponse:
    properties:
      description:
        type: string
      extraAttrs:
        additionalProperties: true
        type: object
      id:
        type: string
      name:
        type: string
      userCount:
        type: integer
    type: object
  types.ListRolesPagedResponse:
    properties:
      page:
        type: integer
      roles:
        items:
          $ref: '#/definitions/types.ListRoleResponse'
        type: array
      size:
        type: integer
      total:
        type: integer
    type: object
...
  types.RolePermissionBindingBody:
    properties:
      action:
        type: string
      resource:
        type: string
      role_id:
        type: string
      scope:
        type: string
    type: object
...
  types.UserRoleBindingBody:
    properties:
      orgId:
        type: string
      roleId:
        type: string
      userId:
        type: string
    required:
    - orgId
    - roleId
    - userId
    type: object
host: localhost:8000
info:
  contact:
    email: jamsheed@nsmail.dev
    name: Jamsheed
    url: http://www.swagger.io/support
  description: This is REST api definitions.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: BigBucks Solutions Auth Engine
  version: 0.0.1
paths:
  /master-data/actions:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - JWTAuth: []
      - JWTAuth: []
      summary: Get actions
      tags:
      - permissions
  /master-data/resources:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - JWTAuth: []
      - JWTAuth: []
      summary: Get resources
      tags:
      - permissions
  /master-data/scopes:
    get:
      consumes:
      - application/json
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
      security:
      - JWTAuth: []
      - JWTAuth: []
      summary: Get scopes
      tags:
      - permissions
...
  /permissions:
    post:
      consumes:
      - application/json
      description: Create a new permission in the system
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Permission object
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/types.CreatePermissionBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - JWTAuth: []
      summary: Create new permission
      tags:
      - permissions
  /roles:
    get:
      consumes:
      - application/json
      description: Get paginated list of roles with user count
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by role name
        in: query
        name: role_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ListRolesPagedResponse'
      security:
      - JWTAuth: []
      summary: List roles
      tags:
      - roles
    post:
      consumes:
      - application/json
      description: Create a new role in the system
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Role object
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/bigbucks_solution_auth_rest-api_controllers_types.Role'
      produces:
      - application/json
      responses:
        "201":
          description: Created
      security:
      - JWTAuth: []
      summary: Create new role
      tags:
      - roles
  /roles/:role_id/permissions:
    post:
      consumes:
      - application/json
      description: Lists permissions of a role
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/types.ListRolePermission'
            type: array
      security:
      - JWTAuth: []
      summary: List permission of a role
      tags:
      - roles
  /roles/bind-permission:
    post:
      consumes:
      - application/json
      description: Associates a permission with a role
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: Binding details
        in: body
        name: rolepermission
        required: true
        schema:
          $ref: '#/definitions/types.RolePermissionBindingBody'
      produces:
      - application/json
      responses:
        "200":
          description: Permission bound successfully
          schema:
            type: string
      security:
      - JWTAuth: []
      summary: Bind permission to role
      tags:
      - roles
  /roles/bind-user:
    post:
      consumes:
      - application/json
      description: Associates a role with a user in an organization
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: User role binding details
        in: body
        name: binding
        required: true
        schema:
          $ref: '#/definitions/types.UserRoleBindingBody'
      produces:
      - application/json
      responses:
        "200":
          description: Role bound to user successfully
          schema:
            type: string
      security:
      - JWTAuth: []
      summary: Bind role to user
      tags:
      - roles
  /roles/unbind-permission:
    post:
      consumes:
      - application/json
      description: Removes a permission with a role
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: UnBinding details
        in: body
        name: rolepermission
        required: true
        schema:
          $ref: '#/definitions/types.RolePermissionBindingBody'
      produces:
      - application/json
      responses:
        "200":
          description: Permission unbound successfully
          schema:
            type: string
      security:
      - JWTAuth: []
      summary: UnBind permission to role
      tags:
      - roles
  /roles/unbind-user:
    post:
      consumes:
      - application/json
      description: Removes a role with a user in an organization
      parameters:
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      - description: User role binding details
        in: body
        name: binding
        required: true
        schema:
          $ref: '#/definitions/types.UserRoleBindingBody'
      produces:
      - application/json
      responses:
        "200":
          description: Role bound to user successfully
          schema:
            type: string
      security:
      - JWTAuth: []
      summary: UnBind role to user
      tags:
      - roles
...
  /user/authorize:
    post:
      consumes:
      - application/json
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.CheckPermissionBody'
      - description: Authorization
        in: header
        name: X-Auth
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.AuthorizeResponse'
        "400":
          description: ""
        "500":
          description: ""
      security:
      - JWTAuth: []
      summary: Check user have permission
      tags:
      - auth
...
securityDefinitions:
  JWTAuth:
    in: header
    name: X-Auth
    type: apiKey
swagger: "2.0"